pyproject.toml
[tool.poetry]
name = "caldav2google"
version = "0.1.0"
description = "A script to convert CalDAV to Google Calendar"
authors = ["Roger Gonzalez <roger@rogs.me>"]
license = "GPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
caldav = "^1.4.0"
icalendar = "^6.1.0"
google-api-python-client = "^2.154.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
python-lsp-ruff = "^2.2.2"
python-lsp-server = "^1.12.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "__init__.py"
]

line-length = 121
indent-width = 4
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "PL", "B", "A", "C4", "TID", "ERA", "RET", "W", "C90", "ARG", "Q", "FLY", "SIM", "COM", "D"]
ignore = ["E402", "PLW2901"]

[tool.ruff.lint.pylint]
max-args = 6

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.pytest.ini_options]
pythonpath = [
    "."
]
